plogging 이동거리 integer-> double로 변경

+++ src/main/java/city/olooe/plogging_project/model/map/PloggingEntity.java
@@ -39,7 +39,7 @@ public class PloggingEntity {
   private String type; // 유형 (제자리시작 / 목표지설정 / 추천경로 / 챌린지경로)
   private Date ploggingTime; // 이동시간
   private Date regDate; // 플로깅 날짜(insert날짜)
-  private Integer distance; // 이동 거리
+  private Double distance; // 이동 거리
   private Boolean status; // 성공여부
 
 }
